+++ types/IntAttribute.ts
@@ -4,7 +4,7 @@
 export class IntAttribute {
     private static type: OpenXmlIo.ReadWriteNamed<{ val: number }> = IntAttribute; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement, expectedTagName: string, parentTags?: string): { val: number } {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element, expectedTagName: string, parentTags?: string): { val: number } {
         xmlDoc.validator.expectNode(elem, expectedTagName, parentTags);
         return {
             val: xmlDoc.attrInt(elem, "val") ?? 0,