+++ types/Point2DType.ts
@@ -2,7 +2,7 @@
 export class Point2DType {
     private static type: OpenXmlIo.ReadWriteNamed<OpenXml.Point2DType> = Point2DType; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement, expectedTagName: string, parentTags?: string): OpenXml.Point2DType {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element, expectedTagName: string, parentTags?: string): OpenXml.Point2DType {
         xmlDoc.validator.expectNode(elem, expectedTagName, parentTags);
         return {
             x: xmlDoc.attrInt(elem, "x") ?? 0,