+++ types/RowId.ts
@@ -3,7 +3,7 @@
 export class RowId {
     private static type: OpenXmlIo.ReadWrite<OpenXml.RowId> = RowId; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement): OpenXml.RowId {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element): OpenXml.RowId {
         return IntElement.read(xmlDoc, elem, "xdr:row", "from, to");
     }