+++ types/RowOffset.ts
@@ -3,7 +3,7 @@
 export class RowOffset {
     private static type: OpenXmlIo.ReadWrite<OpenXml.RowOffset> = RowOffset; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement): OpenXml.RowOffset {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element): OpenXml.RowOffset {
         return IntElement.read(xmlDoc, elem, "xdr:rowOff", "from, to");
     }