+++ types/StringAttribute.ts
@@ -4,7 +4,7 @@
 export class StringAttribute {
     private static type: OpenXmlIo.ReadWriteNamed<{ val: string }> = StringAttribute; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement, expectedTagName: string, parentTags?: string): { val: string } {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element, expectedTagName: string, parentTags?: string): { val: string } {
         xmlDoc.validator.expectNode(elem, expectedTagName, parentTags);
         return {
             val: xmlDoc.attrString(elem, "val") ?? "",