+++ types/StringElement.ts
@@ -4,7 +4,7 @@
 export class StringElement {
     private static type: OpenXmlIo.ReadWriteNamed<{ content: string }> = StringElement; // TODO type-checker
 
-    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: HTMLElement, expectedTagName: string, parentTags?: string): { content: string } {
+    public static read(xmlDoc: OpenXmlIo.ReaderContext, elem: Element, expectedTagName: string, parentTags?: string): { content: string } {
         xmlDoc.validator.expectNode(elem, expectedTagName, parentTags);
         return {
             content: <string><any>elem.textContent, // only null on document or Doctype